Moved the Integrated Circuit libraries to their own directory
